requirements.txt
scikit-learn
pandas
gitpython
fastapi
uvicorn
numpy 